Update .gitignore
The `lintr` package has a lot of underlines for long lines with more than 80 characters, which often happens, so I've customised it using .litr in the root directory and added it to .gitignore. Not sure if I should do this but there doesn't seem to be another way to customise it.