Future changes to trait value substitutions process
See lines 1178-1885 for commented out code. This works for adding substitutions based on individual `words` rather than the entire trait value.

The code works, but causes problems if the same `word` is part of entire strings that have different substitutions.

We need to have an ifelse loop with some sort of a hook indicating whether to use `entire string` or `word` based substitutions.