Fix unit conversions warning (#100)
The unit conversions warning (#65) was showing up again for both @ehwenk and me. It seems to me that there is something going on with the `mutate` `ifelse` statement not being vectorised in the expected way. I don't fully understand why it isn't working as expected but @dfalster suggested a fix (add `value_type` to the `group_by` variables) which seems to produce the right output. (The problem dataset was vanderMoezel_1987.)