generalise `build_update_taxonomy`
A much more simplified version of `build_update_taxonomy` has now been implemented.

This does require a new version of the taxon_list.csv for AusTraits - previously taxon_name was the best-accepted-name and the "notes" were only retained in cleaned_name. Then code mixed-and-matched depending on finding "[" in the name. This obviously doesn't work for a generic workflow.