Update testdata.R
Remove check for units in unit_conversions - this isn't expected. The `unit_in` field can also be a data.csv column. Or there can be units that are used, but happen to not be in the unit_conversions because there aren't any instances where they require converting.

Noted in  issue #146 as well.