Add BibType as a column in methods table (#193)
As described [here](https://github.com/traitecoevo/traits.build/issues/192), the `BibType ` associated with a dataset is used by the R package `RefManageR` to process the references, but is not displayed anywhere within the traits.build database. A column recording the BibType of the
primary citation is now included within the reference table. 
This change has not effected any dataset metadata.yml files, but is
simply capturing and outputting data that was already documented in
them.

Note that `austraits` functions still work and the function
`extract_data` lets you filter data based on `BibType`.