Add identifiers table to traits.build output (#195)
Add a dataframe to a traits.build output that documents "identifiers"
that link specimens/individuals/etc. across datasets and to
museum/herbarium collections.

Six types of identifiers have been added to the traits.build schema,
reflecting key identifiers types specified by DarwinCore. This is not a
comprehensive list of all DarwinCore identifiers - but instead of adding
others, this seems like a good starting point, and traits.build has been
coded to NOT throw an error is a different `identifier_type` is used.

Note, `catalog_number` is for non-universal identifiers that represent
an individual researchers plant tag system while `collectionID` and
`materialSampleID` are meant to be universally unique identifiers.

Notes:
- coded so that metadata.yml files are not required to have an
`identifiers` section, so people already using traits.build do not need
to retrofit their metadata files
- there is a corresponding branch in {austraits} that makes changes to
functions to accommodate the changed output structure. The DESCRIPTION
specifies that {austraits} must be installed from this commit - and will
need to be changed back to master when new releases of {austraits} and
{traits.build} are simultaneously released.
- pull request includes: tests, new function to add identifiers to
metadata files

_I don't understand why changes that were previously pushed to master
are appearing here. For instance, the renaming of functions. Confirming
this pull request is from `add_specimen_ID` branch to `develop`_