fix taxon_name bug
In datasets where all trait values are excluded (in excluded_data), there is a chunk of code in the function `dataset_update_taxonomy` that caused `taxon_name` to become logical and then austraits::bind_databases would break. I'd always thought it was all columns that turned logical and never looked into solving it. But for another user, with 1000's of datasets, some of which are initially empty it was a problem - so fixed.