fix taxon_name bug (#203)
In datasets where all trait values are excluded (in excluded_data),
there is a chunk of code in the function `dataset_update_taxonomy` that
causes `taxon_name` to become logical and then `austraits::bind_databases()`
would break.

This fix has been tested on austraits.build and by Kris Wild and works
as intended.